API Documentation

Docstrings for CALIPSO.jl interface members can be accessed through Julia's built-in documentation system or in the list below.

Contents

API Documentation

Contents

Index

Solver

Trajectory Optimization

Index

CALIPSO.Constraint

CALIPSO.Constraints

CALIPSO.Cost

CALIPSO.Dynamics

CALIPSO.Options

CALIPSO.Solver

CALIPSO.callback_inner

CALIPSO.callback_outer

CALIPSO.empty_constraint

CALIPSO.get_trajectory

CALIPSO.initialize!

CALIPSO.initialize_actions!

CALIPSO.initialize_states!

CALIPSO.linear_interpolation

CALIPSO.solve!

Solver

CALIPSO.Solver — Type

Solver(methods, num_variables, num_parameters, num_equality, num_cone;
    parameters, nonnegative_indices, second_order_indices, custom, options)

CALIPSO solver 

methods: ProblemMethods - includes objective and constraint functions, as we as their derivatives 
num_variables: Int - dimension of primal decision variables 
num_parameters: Int - dimension of problem data 
num_equality: Int - dimension of equality constraints 
num_cone: Int - dimension of cone constraints 
parameters: Vector{Real} - problem data 
nonnegative_indices: Vector{Int} - indices of cone constraints corresponding to nonnegative orthant 
second_order_indices: Vector{Vector{Int}} - indices of cone constraints corresponding to second-order cones
custom: Any - user-provided type used for solver callbacks 
options: Options - solver settings

CALIPSO.solve! — Function

solve!(solver) 

method for optimizing a Solver 

solver: Solver

CALIPSO.initialize! — Function

initialize!(solver, guess) 

method for initializing primal decision variables 

solver: Solver 
guess: Vector{Real} - user-provided initialization for primal decision variables

CALIPSO.Options — Type

Options 

CALIPSO solver settings

CALIPSO.empty_constraint — Function

empty_constraint(x, θ)

convenience method for empty constraints 

x: Vector{Real} - primal decision variables 
θ: Vector{Real} - problem data

CALIPSO.callback_inner — Function

callback_inner(custom, solver)

method called during solver's inner iterations 

custom: Any - user-provided type used for solver callbacks
solver: Solver

CALIPSO.callback_outer — Function

callback_outer(custom, solver)

method called during solver's outer updates 

custom: Any - user-provided type used for solver callbacks
solver: Solver

Trajectory Optimization

CALIPSO.Cost — Type

Cost(cost, num_state, num_action;
    num_parameter, checkbounds, constraint_tensor)

stage-cost type 

cost: Function 
num_state: Int - dimension of state 
num_action: Int - dimension of action 
num_parameter: Int - dimension of problem data 
checkbounds: Bool - flag for checking @inbounds for codegen methods 
constraint_tensor: Bool - flag for generating second-derivative methods

CALIPSO.Constraint — Type

Constraint(constraint, num_state, num_action;
    num_parameter, checkbounds, constraint_tensor)

constraint type 

constraint: Function 
num_state: Int - dimension of state 
num_action: Int - dimension of action 
num_parameter: Int - dimension of problem data 
checkbounds: Bool - flag for checking @inbounds for codegen methods 
constraint_tensor: Bool - flag for generating second-derivative methods

CALIPSO.Constraints — Type

Constraints 

vector of Constraint types

CALIPSO.Dynamics — Type

Dynamics(dynamics, num_next_state, num_state, num_action;
    num_parameter, checkbounds, constraint_tensor)

dynamics type 

dynamics: Function 
num_next_state: Int - dimension of next state
num_state: Int - dimension of current state 
num_action: Int - dimension of current action 
num_parameter: Int - dimension of problem data 
checkbounds: Bool - flag for checking @inbounds for codegen methods 
constraint_tensor: Bool - flag for generating second-derivative methods

CALIPSO.initialize_states! — Function

initialize_states!(solver::Solver, states)

method for initialized primal variables with state trajectory 

solver: Solver 
states: Vector{Vector{Real}} - trajectory of states

CALIPSO.initialize_actions! — Function

initialize_actions!(solver::Solver, actions)

method for initialized primal variables with action trajectory 

solver: Solver 
actions: Vector{Vector{Real}} - trajectory of actions

CALIPSO.get_trajectory — Function

get_trajectory(solver)

method for returning state and action trajectories from solver 

solver: Solver

CALIPSO.linear_interpolation — Function

linear_interpolation(initial_state, final_state, horizon)

method for generating a linear interpolating trajectory 

initial_state: Vector{Real} - first state 
final_state: Vector{Real} - last state 
horizon: Int - length of trajectory